+++ Changes.md
@@ -1,5 +1,15 @@
 
 
+#
+### **+05:30 03:12:03 PM 19-05-2024, Sunday**
+
+  * Added support for any data type on the circular buffer.
+  * The main class `CSE_CircularBuffer` and the functions are now template based.
+  * Renamed `CSE_CircularBuffer.cpp` to `CSE_CircularBuffer.tpp`.
+  * Renamed `CSE_CircularBuffer.h` to `CSE_CircularBuffer.hpp`.
+  * Created new `CSE_CircularBuffer.h` header file for C-style include.
+  * Updated `Test.ino` example.
+  * `Test.ino` example compiles and is working.
 
 #
 ### **+05:30 12:07:47 PM 19-05-2024, Sunday**